model/source/requirements.txt
###### Requirements with Version Specifiers ######
boto3==1.18.47
botocore==1.21.47
s3fs==0.4.2
numpy==1.21.2
pandas==1.3.3
psycopg2-binary==2.9.1
scikit-learn==0.19.2
mlflow==1.23.1
pytest==6.2.5